+++ src/lib.rs
@@ -1,13 +1,14 @@
-pub mod controller;
 pub mod paths;
 pub mod args;
 pub mod opts;
 pub mod urls;
 pub mod feed;
 pub mod feed_item;
-pub mod builder;
 pub mod template;
 pub mod cli;
 pub mod errors;
 pub mod utils;
 pub mod db;
+
+pub mod handlers;
+pub mod builders;