+++ CHANGELOG.md
@@ -6,7 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.0.4] 2024-12-27
+### Fixed
+- Fix parsing age for articles having incorrectly set date (into the future)
 
+### Added
+- Add missing `is_query` field for serialized feeds
+- Add option to override config dir with env var
+
+### Changed
+- Refactored command handlers into separate module
+- Added builders module
+ 
 ## [1.0.3] 2024-12-23
 ### Changed
 - Use crates.io version of libnewsboat